requirements.txt
fastapi[all]
uvicorn[standard]
pymongo
bcrypt
pyjwt
google-generativeai
IPython